Markdown block: renames component
The `MarkdownPreview` and `MarkdownLivePreview` components had very
similar names and that could be confusing. Now `MarkdownPreview `is called
`MarkdownRenderer`.
